Obfuscate RX shenanigans
