+++ app/src/main/java/com/plaid/linksample/MainActivity.kt
@@ -17,16 +17,12 @@ import com.plaid.link.Plaid
 import com.plaid.link.linkTokenConfiguration
 import com.plaid.link.openPlaidLink
 import com.plaid.link.result.PlaidLinkResultHandler
-import com.plaid.linksample.network.LinkSampleApiFactory
-import io.reactivex.Single
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
+import com.plaid.linksample.network.LinkTokenRequester
 
 class MainActivity : AppCompatActivity() {
 
   private lateinit var result: TextView
   private lateinit var tokenResult: TextView
-  private val linkSampleApi = LinkSampleApiFactory.api
 
   private val myPlaidResultHandler by lazy {
     PlaidLinkResultHandler(
@@ -77,33 +73,19 @@ class MainActivity : AppCompatActivity() {
    * [parameter reference](https://plaid.com/docs/link/android/#parameter-reference).
    */
   private fun openLink() {
-    getLinkTokenFromServer().subscribe(
-      {
-        this@MainActivity.openPlaidLink(
-          linkTokenConfiguration = linkTokenConfiguration {
-            token = it
-          }
-        )
-      },
-      { e -> Toast.makeText(this, e.message, Toast.LENGTH_SHORT).show() })
-
+    LinkTokenRequester.token.subscribe(::onLinkTokenSuccess, ::onLinkTokenError)
   }
 
-  /**
-   * In production, make an API request to your server to fetch
-   * a new link_token. Learn more at https://plaid.com/docs/#create-link-token.
-   *
-   * This is a dummy implementation. If you curl for a link_token, you can
-   * copy and paste the link_token value here.
-   */
-  private fun getLinkTokenFromServer(): Single<String> {
-    return linkSampleApi.getLinkToken()
-      .subscribeOn(Schedulers.io())
-      .observeOn(AndroidSchedulers.mainThread())
-      .map { it.link_token }
-    
-    // Optionally, un-comment this to paste your curled link_token here.
-    // return Single.just("<GENERATED_LINK_TOKEN>")
+  private fun onLinkTokenSuccess(linkToken: String) {
+    this@MainActivity.openPlaidLink(
+      linkTokenConfiguration = linkTokenConfiguration {
+        token = linkToken
+      }
+    )
+  }
+
+  private fun onLinkTokenError(error: Throwable) {
+    Toast.makeText(this, error.message, Toast.LENGTH_SHORT).show()
   }
 
   override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
